fix: Resolve 3 critical bugs in valuation engine
Bug Fixes:
1. UnboundLocalError: Added default values for projection parameters
   - revenue_growth, ebit_margin, forecast_years
   - depreciation_pct, capex_pct, wc_change, net_borrowing
   - Prevents crashes when dashboard mode falls back to manual mode

2. StreamlitMixedNumericTypesError: Fixed type mismatch in SGR analysis
   - Wrapped payout_ratio value in float() for consistency
   - Ensures number_input receives same type as min/max values

3. ValueError: Added DataFrame constructor validation
   - Check projections format before creating DataFrame
   - Graceful fallback to empty DataFrame with warning
   - Prevents crashes when projections data is malformed

All fixes ensure graceful degradation rather than hard failures.

+++ analytics/model_inputs_ui.py
@@ -200,7 +200,7 @@ def display_sgr_analysis(financial_data: dict, roe: float,
     with col1:
         payout_ratio = st.number_input(
             "Dividend Payout Ratio (%)",
-            value=sgr_results['payout_ratio'] * 100,
+            value=float(sgr_results['payout_ratio'] * 100),
             min_value=0.0,
             max_value=100.0,
             format="%.1f",
